Made changes in Dice task1
